Move General keybinds to package definitions
In the spirit of keeping everything within the `use-package`
definition, all of the `general` keybinds are defined within the
`use-package` definition for the proper package.

The `window` package is not getting picked up by `use-package`, so
window manipulation binds are  placed in the `emacs` pseudo-package
for now.

There are three other changes included that are not part of migrating
the keybinds, those are:

- Slight clean-up of the config file reload bind
- Remove binds for popup, as they weren't working to begin with.
- Moved Flypsell `LocalWords` to the top of the document