Add function and hook to update dynamic blocks
Function `max/org-update-dblocks` is a wrapper around
`org-update-all-dblocks`. It only runs said function if the major mode
of the current buffer is `org-mode`. The intent is to use "global"
hooks to trigger this function while only effecting Org buffers.

The new before save hook triggers `max/org-update-dblocks` before a
save occurs. This way, the buffer will be saved with updated dynamic
blocks.