Update Capture Note template for dynamic file names
When performing an `org-capture`, the notes template now prompts for a
description, then uses that description to create a new file and
auto-populate a `TITLE` and level one header. The template file
contains a `PROPERTIES` drawer with `CREATED` preset to the current
datetime.

This allows me to quickly capture notes without having to think about
where to put the information. Just a brief description, and then my
notes.