Improve how I am assigning Org directories
Instead of assigning the Org directories in the `init` declaration, I
am assigning them within the `config` declaration. This is because I
do not want Emacs to evaluate the body until after the package is
available.

As well, instead of using a `defvar`, I am using a `let*` statement to
scope the existence of these variables.