New packages and keybindings for Elisp editing
Updated the existing Elisp keybinds into the _eval_ group, and the
_jump_ group.

Added declaration for the IELM package. IELM is REPL for Emacs Lisp.
REPL based development is popular in the Lisp world as it allows for
exploratory programming in a controlled, but temporary, environment. I
need to start trying out more REPL driven development, so I have bound
a key to quickly jump into the IELM REPL.

Added keybindings for Elisp SLIME nav mode. The existing binds fit
more with standard Emacs binds, instead of Evil mode binds.

Add declarations for the Lispy and Lispyville packages. Lispy is
fantastic library for editing Lisp, and Lispyville brings that
Evil goodness.