Enable org-special-ctrl-a/e
`org-special-ctrl-a/e` changes how C-a & C-e work (or "0" and "$" in my case).
Leading stars and any todo keywords are not treated as part of the line, so
going to the beginning of the line starts at the headlines text. Similarly, tags
are not treated as part of the line, so going to the end of the line lands just
before the tags.

Very handy.