Add declaration for Magit on Windows
At work, I have been using Magit more and more for small operations. The three
main use cases have been writing commits, displaying a diff, or showing the
logs. Git CLI on Eshell does not cut it, and opening Git Bash just to show a
diff is annoying. I find it is so far worth the delay to be able to view the
logs in a nice graph, or use Magit's diff display.

Unless I move to Git Bash as my daily shell, I think I will be sticking with
this workflow of Eshell + Magit.