Set fallback fontset to Symbola
If a character is unavailable in my primary font, then Emacs will fall back to
the "fontset-default" font. In my case, the _Symbola_[0] font. _Symbola_ is a
visually clean font that contains most of the characters from the Unicode
standard. The font is unfortunately not under a free license[1].

GNU Unifont[2] would be my first choice, but a bitmap font would clash with my
other font choices and also renders poorly on Windows. GNU also has FreeMono[3],
but that was last updated in 2012.

[0]: https://dn-works.com/ufas/
[1]: https://dn-works.com/wp-content/uploads/2020/UFAS-Docs/License.pdf
[2]: http://unifoundry.com/unifont/index.html