Enable fast, but imprecise scrolling
Fast, but imprecise scrolling helps keep scrolling speed consistent. It does
this by only fontifying the portions of the buffer that will _actually_ be
displayed. That means scrolling to the bottom of a buffer will not fontify
anything but the bottom of the buffer.

Honestly I do not see a huge improvement on GNU/Linux, but on Windows I see an
enormous improvement in speeds.